Refactor: Directly access all core datastructures from here on out
This refactor addresses the "interface" mechanism that was previously in use to avoid being bogged down by variable sized templates everywhere. This system was clunky and annoying to work with, not to mention confusing.

Now, we use type aliased versions of the underlying fixed sized structure, which reduces problems with passing around sizes everywhere. This still comes with its own problems, but should reduce API confusion overall.

WARNING: This may have caused issues with the parser due to a switch to temporary storage over permanent storage in the process of this refactor. However, strings should now be copied directly into their own stored memory in Atlas Sub Images which means this change should have been safe.